requirements.txt
# ─── Core Deep Learning ───────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ─── Data / ML ────────────────────────────────────────────────────────────────
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# ─── Visualisation ────────────────────────────────────────────────────────────
matplotlib>=3.7.0

# ─── Web Interface ────────────────────────────────────────────────────────────
gradio>=4.0.0

# ─── Logging & Experiment Tracking ────────────────────────────────────────────
tensorboard>=2.14.0
